Set the main sections heights and of subsection of header
The main sections are given 25%, 75%, and 5% of the height,
and the flex growth factor is set to 1, 3, and 0 respectively.
The subsection summary-header is given a height of 100%
of its space while its growth is set to 1.

This is because we want the max growth to correspond to these factors
while maintaining the ratio of heights that make content stay within
a set space.