Size and center the action-container
The action container needs to be a similar size width-wise
as the summary-header's contents. These occupy a width of 70%
(35% for the left box and right box respectively) so we set this width.

Additionally, we make the action-call section a flex box and
then center its contents so that action-container is centered with a width of
70%